Initial work to support classes as custom properties
So far all of the changes are just on the data level, no UI changes have
been made yet. Change is WIP since it currently leaves the Enums Editor
broken.

It appears to me that a ClassPropertyType will have some overlap with
the already supported ObjectType. I think it would be nice if we could
merge these.